refactor utils::config
